add middlewares to routes

+++ routes/api.php
@@ -4,6 +4,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BookController;
 use App\Http\Controllers\CustomerController;
+use App\Http\Controllers\ApiAuthController;
+use App\Models\User;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,21 +19,41 @@ use App\Http\Controllers\CustomerController;
 */
 
 Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
-    return $request->user();
+    return User::with('customer')->find($request->user()->id);
 });
 
 Route::group(['prefix' => 'book'], function () {
     Route::get('/', [BookController::class, 'index']);
     Route::get('/search', [BookController::class, 'search']);
-    Route::post('/', [BookController::class, 'store']);
+    Route::middleware(['admin'])->post('/', [BookController::class, 'store']);
     Route::get('/{id}', [BookController::class, 'show']);
-    Route::put('/{id}', [BookController::class, 'update']);
-    Route::delete('/{id}', [BookController::class, 'destroy']);
+    Route::middleware(['admin'])->put('/{id}', [BookController::class, 'update']);
+    Route::middleware(['admin'])->delete('/{id}', [BookController::class, 'destroy']);
 });
 
 Route::group(['prefix' => 'customer'], function () {
-    Route::get('/', [CustomerController::class, 'index']);
-    Route::get('/{id}', [CustomerController::class, 'show']);
-    Route::put('/{id}', [CustomerController::class, 'update']);
-    Route::delete('/{id}', [CustomerController::class, 'destroy']);
+    Route::middleware(['admin'])->get('/', [CustomerController::class, 'index']);
+    Route::middleware(['auth:sanctum'])->get('/{id}', [CustomerController::class, 'show']);
+    Route::middleware(['auth:sanctum'])->put('/{id}', [CustomerController::class, 'update']);
+    Route::middleware(['admin'])->delete('/{id}', [CustomerController::class, 'destroy']);
 });
+
+Route::group(['prefix' => 'auth'], function () {
+    Route::post('/register', [ApiAuthController::class, 'register']);
+    Route::post('/login', [ApiAuthController::class, 'login']);
+    Route::middleware(['auth:sanctum'])->post('/logout', [ApiAuthController::class, 'logout']);
+});
+
+Route::get('unauthenticated', function () {
+    return response()->json([
+        'message' => 'unauthenticated',
+        'endpoints' => [
+            ['endpoint' => 'api/auth/login', 'method' => 'POST', 'shape' => [
+                'email', 'password'
+            ]], ['endpoint' => 'api/auth/register', 'method' => 'POST', 'shape' => [
+                'name', 'email', 'password', 'address_line_1', 'address_line_2', 'phone_number', 'city', 'parish'
+            ]],
+            ['endpoint' => 'api/auth/logout', 'method' => 'POST', 'shape' => []]
+        ]
+    ], 401);
+})->name('unauthenticated');
